requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
python-frontmatter>=1.1.0
rich>=13.0.0
typer>=0.9.0
pyyaml>=6.0
pytest>=7.0.0
